requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
numpy==1.24.4
psycopg2==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
pyshp==2.3.1
sqlparse==0.5.1
tzdata==2024.1